refactor: merge punctuation checks into typography.punctuation
