--- proselint/checks/hyperbole.py
@@ -1,29 +1,0 @@
-"""
-Hyperbolic language.
-
----
-layout:     post
-source:     ???
-source_url: ???
-title:      hyperbolic language
-date:       2014-06-10 12:31:19
-categories: writing
----
-
-Hyperbolic language.
-
-"""
-from proselint.tools import existence_check
-
-
-def check(text):
-    """Check the text."""
-    err = "hyperbole.misc"
-    msg = "'{}' is hyperbolic."
-
-    words = [
-        r"[a-z]*[!]{2,}",
-        r"[a-z]*\?{2,}"
-    ]
-
-    return existence_check(text, words, err, msg)
